+++ modules/web/main.tf
@@ -92,19 +92,19 @@ resource "aws_ecs_task_definition" "web_task" {
       },
       {
         name  = "CF_API_URL"
-        value = var.app_url
+        value = coalesce(var.controlplane_api_url, var.app_url)
       },
       {
         name  = "CF_ACCESS_API_URL"
-        value = var.app_url
+        value = coalesce(var.access_api_url, var.app_url)
       },
       {
         name  = "CF_AUTHZ_URL",
-        value = var.app_url
+        value = coalesce(var.authz_api_url, var.app_url)
       },
       {
         name  = "CF_AUTHZ_GRAPH_URL",
-        value = "${var.app_url}/graph"
+        value = "${coalesce(var.authz_api_url, var.app_url)}/graph"
       },
       {
         name  = "CF_TEAM_NAME"
@@ -169,7 +169,7 @@ resource "aws_ecs_service" "web_service" {
 }
 resource "aws_lb_listener_rule" "service_rule" {
   listener_arn = var.alb_listener_arn
-  priority     = 100
+  priority     = var.alb_listener_rule_priority
   action {
     type             = "forward"
     target_group_arn = aws_lb_target_group.web_tg.arn