+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # @common-fate/terraform-aws-common-fate-deployment
 
+## 1.10.0
+
+### Minor Changes
+
+- 1ac549b: IAM Identity Center integration: adds optional permissions to allow management of group memberships
+- 8184a23: Adds variables to customise API URLs and CORS allowed origins
+
+### Patch Changes
+
+- 33efd74: Fixes an issue where the default logo URL pointed to a broken image if a custom logo wasn't provided for the deployment
+
+## 1.9.0
+
+### Minor Changes
+
+- 8bfad52: Add rds deletion protection and remove old security groups
+
 ## 1.8.0
 
 ### Minor Changes